requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django_filter==21.1
django_extensions==3.1.5
django-environ==0.8.1
gunicorn==20.1.0
ipython==7.29.0
psycopg2==2.9.2
redis==3.5.3
